Secure 'clean' command
I was using 'rm -rf ...' command and as I do not trust how variables
behave, I could not risk it keeping the recursive and force flags in 'rm'
command.

You never know how, for example, "rm -rf $(OBJDIR)" it might behave *IF*
our $(OBJDIR) is set up in such way that points to a different directory than
the one we have set ourselves.

Therefore, what I did was to remove the two by adding wildcard for object and
binary files *and* run a separate command for removing the empty folder(s)
with the help of 'rmdir'.

In other words:
    @rm $(OBJDIR)/*.o $(BINDIR)/*
    @rmdir $(OBJDIR) $(BINDIR)

That's a lot better and safer, at least to my eyes.

Oh, and I have secured it with "ifneq" condition to avoid producing
error message if I repeat the 'make clean' command multiple times.

If both folders are not empty, delete their contents and then delete the
folders themselves.