Project refactored
I have decided to simplify mproj, because I was really mind numb to
realize one obvious thing: with every new release of GCC, I would be
forced to add a new set of standards to the project and that would be an
overkill to bloat what started to be a lightweight project generator.

Therefore, I have removed all C and C++ samples, but left a single
sample for each language and use that instead.

If anyone wants to use a different `-std=' flag, he or she can change it
accordingly; as simple as that!