Important improvements
Special thanks to Jefferson Gonzalez (@jgmdev) for finding an important
bug with ccache.

I forgot to validate ccache's existence first and then use it.

It was breaking the basename mechanism and I had to add a new variable
that checks whether ccache is installed or not.

Now it works as expected.

Also, I have discovered that make comes with a ``MAKEFLAGS'' variable
that can be set to the number of jobs you want to execute at the same time.

Again, I would like to thank Jeff for bringing up this necessary
mechanism, which lead me to discover the aforementioned flag.