Some innocent, yet important tweaks
If for some reason contents of 'bin' or 'obj' folders get deleted, the
'make clean' command would complain about not finding the necessary
files for deletion.

I added '-f' flag to force the deletion by bypassing the complaining.

Also, the important change was to remove 'all' from make command and
added it to 'full' by first executing 'build' and then adding $(TARGET).

This way, the user can run 'make full' and succeed in building the
project without complains from make that could not find the necessary
directories to place the generated object and target files.